+++ src/utils/log.mli
@@ -1,8 +1,9 @@
 include Logs.LOG
 
+val is_bench_enabled : unit -> bool
+
 val bench_fn : string -> (unit -> 'a) -> 'a
 
 val bench : 'a Logs.log
 
-val setup :
-  Fmt.style_renderer option -> Logs.level option -> with_timings:bool -> unit
+val setup : Fmt.style_renderer option -> Logs.level option -> bench:bool -> unit